requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
playwright==1.58.0
redis==5.2.1
arq==0.26.1
aiosqlite==0.20.0
markdownify==0.14.1
readability-lxml==0.8.1
lxml==5.3.0
lxml_html_clean==0.4.1
beautifulsoup4==4.12.3
google-genai==1.14.0
tenacity==9.0.0
structlog==24.4.0
playwright-stealth==1.0.6
setuptools<81

# V3: cache + routing + PDF + memoization
msgpack==1.1.0
zstandard==0.23.0
tldextract==5.1.3
feedparser==6.0.11
PyGithub==2.5.0
arxiv==2.1.3
pymupdf==1.25.1
pdfplumber==0.11.6
# asyncpraw is optional — only needed if REDDIT_CLIENT_ID is configured.
# 7.8.1 pins aiosqlite<=0.17.0 which conflicts with our 0.20.0.
# Install manually into a separate venv if/when PRAW route is needed:
#   pip install 'asyncpraw>=8.0.0' (once released) OR pin aiosqlite==0.17.0
# The reddit_praw route has a try/except ImportError guard — service runs
# fine without asyncpraw installed, the route just no-ops.

pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0
respx==0.22.0
